[Task2.3a] Moved fake 'Plant' class used in testing to the app; using it as a baseline for our Demo.
