+++ app/CMakeLists.txt
@@ -10,7 +10,15 @@ set(CMAKE_INCLUDE_CURRENT_DIR ON)
 find_package(Qt5 COMPONENTS Core Widgets Charts REQUIRED)
 find_package(QCustomPlot REQUIRED)
 
-add_executable(demo demo.cpp Controller.cpp Limits.cpp PID.cpp Model.cpp demo/window.cpp)
+set(CPP_SOURCES
+  demo.cpp
+  Controller.cpp
+  Limits.cpp PID.cpp
+  Model.cpp
+  demo/window.cpp
+  fake/plant.cpp)
+
+add_executable(demo ${CPP_SOURCES})
 
 target_link_libraries(demo Threads::Threads Qt5::Widgets Qt5::Charts)
 include_directories(
